modify triangular solver
