+++ src/solver/triangular/mc.cpp
@@ -1,0 +1,14 @@
+#include "dlaf/solver/triangular/mc.h"
+
+namespace dlaf {
+namespace solver {
+namespace internal {
+
+DLAF_SOLVER_TRIANGULAR_MC_ETI(, float)
+DLAF_SOLVER_TRIANGULAR_MC_ETI(, double)
+DLAF_SOLVER_TRIANGULAR_MC_ETI(, std::complex<float>)
+DLAF_SOLVER_TRIANGULAR_MC_ETI(, std::complex<double>)
+
+}
+}
+}
